Changed region method to new class per command method
